+++ CHANGELOG.rst
@@ -55,3 +55,5 @@ Fixed
   Store
 - ``rasa nlu test`` doesn't error anymore when a test file is passed with ``-u``
 - in interactive learning: only updates entity values if user changes annotation
+- ``SQLTrackerStore.keys()`` now returns the distinct stored sender ids
+  instead of the column names